address various warnings from MSVC, float conversion and so on (#383)
* address various warnings from MSVC, float conversion and so on

+++ IGraphics/Controls/IVMultiSliderControl.h
@@ -93,7 +93,7 @@ public:
 
     if(direction == EDirection::Vertical)
     {
-      value = 1. - (y-bounds.T) / bounds.H();
+      value = 1.f - (y-bounds.T) / bounds.H();
       
       for(auto i = 0; i < nVals; i++)
       {
@@ -123,7 +123,7 @@ public:
     if (sliderTest > -1)
     {
       SetValue(mMinTrackValue + Clip(value, 0.f, 1.f) * (mMaxTrackValue - mMinTrackValue), sliderTest);
-      OnNewValue(sliderTest, GetValue(sliderTest));
+      OnNewValue(sliderTest, static_cast<float>(GetValue(sliderTest)));
 
       mSliderHit = sliderTest;
 
@@ -148,7 +148,7 @@ public:
           {
             float frac = (float)(i - lowBounds) / float(highBounds-lowBounds);
             SetValue(linearInterp(GetValue(lowBounds), GetValue(highBounds), frac), i);
-            OnNewValue(i, GetValue(i));
+            OnNewValue(i, static_cast<float>(GetValue(i)));
           }
         }
       }
